mtd: nand: sunxi: implement specific ->read_oob()/->write_oob() functions
Allwinner's ECC engine is capable of protecting a few byte of the OOB area.
Implement specific OOB functions to benefit from this capability.

Also, when in raw mode, the randomizer is disabled, which means you'll only
be able to retrieve randomized data, which is not really useful for most
applications.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>