UBIFS: force LEB atomic updated in ubifs_leb_unmap()
Some places in UBIFS are lazily calling ubifs_leb_unmap() without
calling ubifs_leb_write() or ubifs_leb_map(), thus leaving potential
older versions of the LEB around.

This should not be a problem if UBIFS keep track of all the valid LEBs,
but if there's a bug in the LEB unmapping/unreferencing sequence and
a power-cut occurs, we can then point to a completely outdated version
of the LEB.

Force LEB atomic update with zero len in ubifs_leb_unmap() to force
pushing an empty version of the LEB, and turn ubifs_leb_map() into a
NOP when the LEB is already mapped.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>